adjust data type of facebook id in users table

+++ db/migrate/20141128193928_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration
     create_table :users do |t|
       t.string :first_name
       t.string :last_name
-      t.bigint :facebook_id
+      t.integer :facebook_id
       t.string :email
       t.string :access_key
